move shared logic to shared module
