+++ etl/steps/data/garden/covid/latest/cases_deaths.py
@@ -1,10 +1,9 @@
 """Load a meadow dataset and create a garden dataset."""
 
-from typing import cast
-
 import numpy as np
 import pandas as pd
 from owid.catalog import Dataset, Table
+from shared import add_population_daily, fill_date_gaps
 
 from etl.data_helpers import geo
 from etl.helpers import PathFinder, create_dataset
@@ -74,7 +73,14 @@ def run(dest_dir: str) -> None:
     ## Drop rows
     tb = discard_rows(tb)
     ## Add population
-    tb = add_population_daily(tb, ds_population)
+    tb = add_population_daily(
+        tb,
+        ds_population,
+        missing_countries={
+            "International",
+            "Pitcairn",
+        },
+    )
     ## Add regions
     tb = add_regions(tb, ds_regions, ds_income)
     ## Add period-aggregtes, doublig days
@@ -156,33 +162,6 @@ def aggregate_international(tb: Table) -> Table:
     assert countries_duplicate == {"International"}, "There are unexpected duplicates!"
     # Aggregate
     tb = tb.groupby(["country", "date"]).sum(min_count=1).reset_index()  # type: ignore
-    return tb
-
-
-def fill_date_gaps(tb: Table) -> Table:
-    """Ensure dataframe has all dates.
-
-    Apparently, in the past the input data had all the dates from start to end.
-
-    Early in 2024 this stopped to be like this, maybe due to a change in how the data is reported by the WHO. Hence, we need to make sure that there are no gaps!
-    Source of change might be this: https://github.com/owid/covid-19-data/commit/ed73e7113344caffc9e445946979e1964720348b#diff-cb6c8f3daa43ff50c0cac819d63ce03bedfd4c7cf98ace02cad543a485c9513e
-
-    We do this by:
-        - Reindexing the dataframe to have all dates for all locations.
-    """
-    # Ensure date is of type date
-    tb["date"] = pd.to_datetime(tb["date"], format="%Y-%m-%d").astype("datetime64[ns]")
-
-    # Get set of locations
-    countries = set(tb["country"])
-    # Create index based on all locations and all dates
-    complete_dates = pd.date_range(tb["date"].min(), tb["date"].max())
-
-    # Reindex
-    tb = tb.set_index(["country", "date"])
-    new_index = pd.MultiIndex.from_product([countries, complete_dates], names=["country", "date"])
-    tb = tb.reindex(new_index).sort_index().reset_index()
-
     return tb
 
 
@@ -509,45 +488,3 @@ def set_dtypes(tb: Table) -> Table:
         },
     }
     return tb.astype(dtypes)
-
-
-def add_population_daily(tb: Table, ds_population: Dataset) -> Table:
-    """Add `population` column to table.
-
-    Adds population value on a daily basis (extrapolated from yearly data).
-    """
-    countries_start = set(tb["country"].unique())
-    tb_pop = make_table_population_daily(
-        ds_population=ds_population, year_min=tb["date"].dt.year.min() - 1, year_max=tb["date"].dt.year.max() + 1
-    )
-    tb = tb.merge(tb_pop, on=["country", "date"])
-    countries_end = set(tb["country"].unique())
-
-    # Check countries that went missing
-    countries_missing = countries_start - countries_end
-    assert countries_missing == {
-        "International",
-        "Pitcairn",
-    }, f"Missing countries don't match the expected! {countries_missing}"
-
-    return tb
-
-
-def make_table_population_daily(ds_population: Dataset, year_min: int, year_max: int) -> Table:
-    """Create table with daily population.
-
-    Uses linear interpolation.
-    """
-    # Load population table
-    population = ds_population.read_table("population")
-    # Filter only years of interest
-    population = population[(population["year"] >= year_min) & (population["year"] <= year_max)]
-    # Create date column
-    population["date"] = pd.to_datetime(population["year"].astype("string") + "-07-01")
-    # Keep relevant columns
-    population = population.loc[:, ["date", "country", "population"]]
-    # Add missing dates
-    population = fill_date_gaps(population)
-    # Linearly interpolate NaNs
-    population = geo.interpolate_table(population, "country", "date")
-    return cast(Table, population)
